Correct collection rate, add scrap scenario capacity
adding scenario names in the form described in the Integration or Sensitivity classes allows running of scrap supply and demand scenarios, with Integration.hyperparam containing reasonably detailed descriptions of their workings. There was an issue with collection rates that caused an imbalance in scrap supply and demand at the start of each simulation; this has been fixed. Remaining thing that could be done: in allowing for (scrap) trade, we could allow a more global collection rate setting method that would give an implicit initial scrap trade, rather than the current regional discrepancy.