Create Individual.py
Added file for the Individual class, which takes in the filename and relevant characteristics of a model run output pickle file, and makes the data available in more useful forms, and can also plot the results in various forms. I would recommend just copying and pasting this class into whatever notebook you are using, so that you can make updates to it as needed.