More bug fixes
If any case study data columns were empty but had the column header, it was taking those nan values and trying to put them places. That has been addressed. There may have been additional issues in refining that were fixed in identifying that as the root of this error, but they might not have ever actually been relevant, woops