Alternative Bayes, alt volume drivers for gold
Alternative Bayesian method:
- allows use of thresholds, with way more work than I'd realized going into it. Hopefully it improves performance, can be disabled by setting using_thresholds to False in Sensitivity initialization

Gold volume drivers:
Realized this could be the problem, that we were using industrial to represent bar and coin, transport as jewelry. These sectors have not experienced the same volume growth rates, and by putting them in (for gold only) we can address the issues we've been having with demand, where the model cannot account for the plateau.
- can disable by setting the use_alternative_gold_volumes to false in initialization of Sensitivity class.
- If running a standard Integration(), need to set hyperparam['commodity'] to Au