Changed metrics from sums to average
It shouldn't change anything since it is just scaling by a factor (which when logged means it's minus some number). Also changed from sum of logs to log of sum. Again it probably doesn't change much but I realised that I had tested model performance with log of sum but the code had sum of log. Log of sum should be a nicer function as it doesn't plunge to minus infinity as often.