Adding mining tuning
added a bit to the historical_sim_check_demand() function that lets it run the optimization on mine production and hyperparameters, and it seems to have helped quite a bit on gold. Then, we can leave out several of the mining-related hyperparameters and I've added functionality such that we can still tune some of the remaining hyperparameters (since in the historical tuning I couldn't account for TCRC, but their tuning ranges are much smaller.

Other changes are just background for fixing random things I found problems with for gold, etc.