Decreasing file size
also a few minor things, but the big change is that now the historical data outside the scenario running period won't get saved, nor with the mine-level data. If you want the historical data, you'll have to run the scenario separately or so something with the parameters trim_result_df and save_mining_info entering the senstivitiy function.